Plot the ECDF-based plots.

See vignette("rank_visualizations") for more details. See the methods for data_for_ecdf_plots() for available data formats.

plot_ecdf(
  x,
  variables = NULL,
  K = NULL,
  gamma = NULL,
  prob = 0.95,
  size = 1,
  alpha = 0.33,
  combine_variables = NULL,
  ecdf_alpha = NULL,
  ...,
  parameters = NULL
)

plot_ecdf_diff(
  x,
  variables = NULL,
  K = NULL,
  gamma = NULL,
  prob = 0.95,
  size = 1,
  alpha = 0.33,
  combine_variables = NULL,
  ecdf_alpha = NULL,
  ...,
  parameters = NULL
)

Arguments

x

object supporting the data_for_ecdf_plots() method.

variables

optional subset of variables to show in the plot

K

number of uniformly spaced evaluation points for the ECDF or ECDFs. Affects the granularity of the plot and can significantly speed up the computation of the simultaneous confidence bands. Defaults to the smaller of number of ranks per variable and the maximum rank.

gamma

TODO

prob

the width of the plotted confidence interval for the ECDF.

size

size passed to ggplot2::geom_ribbon() for the confidence band

alpha

alpha level of the confidence band

combine_variables

optionally specify a named list where each entry is a character vectors which specifies a group of variables that will be displayed in the same panel. Panel title will be the name of the list element. A function that takes a character vector as an input and produces a list can also be specified (see combine-functions).

ecdf_alpha

the alpha level of the empirical CDF. Can be either a single number or a function taking the number of variables that were combined (when combine_variables is specified) and returns a number. By default, plots showing many ECDFs will have reduced alpha.

...

additional arguments passed to data_for_ecdf_plots(). Most notably, if x is matrix, a max_rank parameter needs to be given.

parameters

DEPRECATED, use variables instead.

Details

arxiv::1903.08008 by A. Vehtari et al.

See also