Combine an example backend with an example generator to provide full results that can be used to test other functions in the package.

Except for example = "visualizations", all examples will actually compile and fit Stan models and thus may take a while to complete.

SBC_example_results(
  example = c("normal_ok", "normal_bad", "visualizations"),
  interface = c("rstan", "cmdstanr", "rjags"),
  N = 100,
  n_sims = 50
)

Arguments

example

name of the example. normal_ok is an example where the generator matches the model (using the normal generator and normal_sd backend), while normal_bad is an example with a mismatch between the generator and backend that manifests in SBC (normal_bad combines the normal generator with normal_bad backend). visualizations creates a purely artificial results that are meant to showcase the built-in plots (the interface parameter will be ignored).

interface

name of the interface to be used for the backend

N

number of datapoints to simulate from the generator for each simulation

n_sims

number of simulations to perform