Provide a file where a cached result of a fit will be/was stored.

This is to allow you to directly access the cache, should you need to. For normal fits, the cache file will always be a list readable by readRDS and have element $fit for the actual fit. Other elements will contain the text output, messages and warnings emitted by the fit.

cached_fit_filename(cache_dir, backend, generated, suffix = "")

Details

This is also used internally to cache some extra data (notably bridgesampling results) using a non-empty suffix string.