Plot the simulated "true" values versus posterior estimates

Plot the simulated "true" values versus posterior estimates

plot_sim_estimated(
  x,
  variables = NULL,
  estimate = "mean",
  uncertainty = c("q5", "q95"),
  alpha = NULL,
  parameters = NULL
)

Arguments

x

object containing results (a data.frame or SBC_results() object).

variables

variables to show in the plot or NULL to show all

estimate

which estimate to use for the central tendency, must correspond a field already computed in the results (most likely "mean" and "median").

uncertainty

which estimates to use for uncertainty (a character vector of length 2) must correspond a field already computed in the results. Pass NULL to avoid showing uncertainty at all.

alpha

the alpha for the points and uncertainty intervals

parameters

DEPRECATED, use variables instead

Value

a ggplot2 plot object